Additional file renaming.
