+++ framework/math/quadratures/angular/angular_quadrature.cc
@@ -1,4 +1,4 @@
-#include "framework/math/quadratures/angular/angular_quadrature_base.h"
+#include "framework/math/quadratures/angular/angular_quadrature.h"
 #include "framework/math/quadratures/angular/legendre_poly/legendrepoly.h"
 #include "framework/logging/log.h"
 #include "framework/runtime.h"
