--- framework/math/quadratures/angular/angular_quadrature_prodcust.cc
@@ -1,1 +1,0 @@
-#include "framework/math/quadratures/angular/angular_product_quadrature.h"
