+++ framework/math/quadratures/angular/product_quadrature.cc
@@ -1,6 +1,6 @@
-#include "framework/math/quadratures/angular/angular_product_quadrature.h"
-#include "framework/math/quadratures/quadrature_gausslegendre.h"
-#include "framework/math/quadratures/quadrature_gausschebyshev.h"
+#include "framework/math/quadratures/angular/product_quadrature.h"
+#include "framework/math/quadratures/gausslegendre_quadrature.h"
+#include "framework/math/quadratures/gausschebyshev_quadrature.h"
 #include "framework/logging/log.h"
 #include "framework/runtime.h"
 #include <cmath>
