+++ framework/math/quadratures/angular/spherical_quadrature.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "framework/math/quadratures/angular/curvilinear_angular_quadrature.h"
+#include "framework/math/quadratures/angular/curvilinear_quadrature.h"
 #include "framework/math/quadratures/spatial/spatial_quadrature.h"
 
 namespace opensn
