+++ framework/math/spatial_discretization/cell_mappings/finite_element/piecewise_linear/piecewise_linear_polygon_mapping.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "framework/math/spatial_discretization/cell_mappings/finite_element/piecewise_linear/piecewise_linear_base_mapping.h"
-#include "framework/math/quadratures/spatial/quadrature_line.h"
-#include "framework/math/quadratures/spatial/quadrature_triangle.h"
+#include "framework/math/quadratures/spatial/line_quadrature.h"
+#include "framework/math/quadratures/spatial/triangle_quadrature.h"
 #include "framework/mesh/cell/cell.h"
 #include <array>
