+++ lua/framework/math/quadratures/create_curvilinear_product_quadrature.cc
@@ -1,9 +1,9 @@
 #include "framework/lua.h"
 #include "quadratures_lua.h"
-#include "framework/math/quadratures/quadrature_gausschebyshev.h"
-#include "framework/math/quadratures/quadrature_gausslegendre.h"
-#include "framework/math/quadratures/angular/cylindrical_angular_quadrature.h"
-#include "framework/math/quadratures/angular/spherical_angular_quadrature.h"
+#include "framework/math/quadratures/gausschebyshev_quadrature.h"
+#include "framework/math/quadratures/gausslegendre_quadrature.h"
+#include "framework/math/quadratures/angular/cylindrical_quadrature.h"
+#include "framework/math/quadratures/angular/spherical_quadrature.h"
 #include "framework/logging/log.h"
 #include "framework/console/console.h"
 #include "framework/runtime.h"
