+++ lua/framework/math/quadratures/get_prod_quadrature.cc
@@ -1,6 +1,6 @@
 #include "framework/lua.h"
 #include "quadratures_lua.h"
-#include "framework/math/quadratures/angular/angular_product_quadrature.h"
+#include "framework/math/quadratures/angular/product_quadrature.h"
 #include "framework/logging/log.h"
 #include "framework/console/console.h"
 #include "framework/runtime.h"
