+++ lua/framework/math/quadratures/sldfesq/create_sldfesq_quadrature.cc
@@ -1,5 +1,5 @@
 #include "framework/lua.h"
-#include "framework/math/quadratures/angular/sldfe_sq.h"
+#include "framework/math/quadratures/angular/sldfe_sq_quadrature.h"
 #include "framework/console/console.h"
 #include "framework/runtime.h"
 #include "sldfe_lua.h"
