+++ lua/framework/math/quadratures/sldfesq/locally_refine_sldfesq.cc
@@ -1,5 +1,5 @@
 #include "framework/lua.h"
-#include "framework/math/quadratures/angular/sldfe_sq.h"
+#include "framework/math/quadratures/angular/sldfe_sq_quadrature.h"
 #include "framework/console/console.h"
 #include "framework/logging/log.h"
 #include "framework/runtime.h"
