+++ modules/linear_boltzmann_solvers/discrete_ordinates_solver/lbs_discrete_ordinates_solver.cc
@@ -13,8 +13,7 @@
 #include "modules/linear_boltzmann_solvers/lbs_solver/source_functions/source_function.h"
 #include "modules/linear_boltzmann_solvers/lbs_solver/groupset/lbs_groupset.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
-#include "framework/math/quadratures/angular/angular_quadrature_base.h"
-#include "framework/math/quadratures/angular/angular_product_quadrature.h"
+#include "framework/math/quadratures/angular/product_quadrature.h"
 #include "framework/logging/log.h"
 #include "framework/logging/log_exceptions.h"
 #include "framework/utils/timer.h"
