+++ modules/linear_boltzmann_solvers/lbs_solver/groupset/lbs_groupset.h
@@ -6,7 +6,7 @@
 #include "modules/linear_boltzmann_solvers/lbs_solver/groupset/lbs_group.h"
 #include "modules/linear_boltzmann_solvers/lbs_solver/lbs_structs.h"
 #include "framework/math/quadratures/angular/legendre_poly/legendrepoly.h"
-#include "framework/math/quadratures/angular/angular_quadrature_base.h"
+#include "framework/math/quadratures/angular/angular_quadrature.h"
 #include "framework/math/unknown_manager/unknown_manager.h"
 #include "framework/physics/physics_namespace.h"
 #include "framework/utils/utils.h"
